fix: rename constants.ts to index.ts
