+++ practice/e-commerce-website/src/components/Button/Button.tsx
@@ -1,6 +1,6 @@
 /* Import constants */
-import { BUTTON_VARIANTS } from '../../constants/constants';
-import { COMPONENT_SIZES } from '../../constants/constants';
+import { BUTTON_VARIANTS } from '../../constants';
+import { COMPONENT_SIZES } from '../../constants';
 /*Import CSS*/
 import './Button.css';
