+++ practice/e-commerce-website/src/components/Price/Price.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 
 import Price from './Price';
 
-import { COMPONENT_SIZES } from '../../constants/constants';
+import { COMPONENT_SIZES } from '../../constants';
 
 const meta: Meta<typeof Price> = {
   component: Price
