+++ practice/e-commerce-website/src/components/Rating/Rating.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 /* Import constants */
-import { COMPONENT_SIZES } from '../../constants/constants';
+import { COMPONENT_SIZES } from '../../constants';
 /*Import CSS*/
 import './Rating.css';
