+++ practice/e-commerce-website/src/constants/index.ts
