+++ practice/e-commerce-website/src/services/ProductService.ts
@@ -1,7 +1,7 @@
 import { HttpService } from './HttpService';
 import { QueryParams } from '../types/QueryParams';
 import { Product } from '../types/Procduct';
-import { PRODUCTS_API } from '../constants/constants';
+import { PRODUCTS_API } from '../constants';
 
 export class ProductService {
   service = new HttpService(PRODUCTS_API);
